+++ src/app/core/service/index.ts
@@ -2,6 +2,7 @@ export * from './app-translate.service';
 export * from './app.service';
 export * from './browser.service';
 export * from './game.service';
+export * from './logger.service';
 export * from './renderer.service';
 export * from './storage.service';
 export * from './window.service';