+++ src/app/data/poe/service/trade-http.service.ts
@@ -4,12 +4,12 @@ import { BrowserService } from '@app/service';
 import { environment } from '@env/environment';
 import { Language } from '@shared/module/poe/type';
 import { Observable, of, throwError } from 'rxjs';
-import { delay, flatMap, map, retry, retryWhen } from 'rxjs/operators';
+import { delay, flatMap, map, retryWhen } from 'rxjs/operators';
 import { TradeFetchResult, TradeItemsResult, TradeLeaguesResult, TradeResponse, TradeSearchRequest, TradeSearchResponse, TradeStaticResult, TradeStatsResult } from '../schema/trade';
 
 const RETRY_COUNT = 3;
 const RETRY_DELAY = 100;
-const RETRY_LIMIT_DELAY = 100;
+const RETRY_LIMIT_DELAY = 300;
 
 @Injectable({
     providedIn: 'root'
@@ -44,11 +44,16 @@ export class TradeHttpService {
         return this.http.post(url, request, {
             responseType: 'text',
             observe: 'response'
-        }).pipe(retry(RETRY_COUNT), map(response => {
-            const result = JSON.parse(response.body) as TradeSearchResponse;
-            result.url = `${url.replace('/api', '')}/${encodeURIComponent(result.id)}`;
-            return result;
-        }));
+        }).pipe(
+            retryWhen(errors => errors.pipe(
+                flatMap((response, count) => this.handleError(url, response, count))
+            )),
+            map(response => {
+                const result = JSON.parse(response.body) as TradeSearchResponse;
+                result.url = `${url.replace('/api', '')}/${encodeURIComponent(result.id)}`;
+                return result;
+            })
+        );
     }
 
     public fetch(itemIds: string[], queryId: string, language: Language): Observable<TradeResponse<TradeFetchResult>> {
@@ -60,18 +65,29 @@ export class TradeHttpService {
                 }
             })
         }).pipe(retryWhen(errors => errors.pipe(
-            flatMap((error: HttpErrorResponse) => {
-                if (error.status === 403) {
-                    return this.browser.retrieve(url).pipe(
-                        map(() => null)
-                    );
-                }
-                if (error.status === 429) {
-                    return of(error).pipe(delay(RETRY_LIMIT_DELAY));
-                }
-                return throwError(error);
-            })
+            flatMap((response, count) => this.handleError(url, response, count))
         )));
+    }
+
+    private getAndTransform<TResponse>(url: string): Observable<TResponse> {
+        return this.http.get(url, {
+            responseType: 'text',
+            observe: 'response'
+        }).pipe(
+            retryWhen(errors => errors.pipe(
+                flatMap((response, count) => this.handleError(url, response, count))
+            )),
+            map(response => this.transformResponse(response))
+        );
+    }
+
+    private transformResponse<TResponse>(response: HttpResponse<string>): TResponse {
+        const result = response.body.replace(
+            /\\u[\dA-Fa-f]{4}/g,
+            (match) => String.fromCharCode(
+                parseInt(match.replace(/\\u/g, ''), 16))
+        );
+        return JSON.parse(result) as TResponse;
     }
 
     private getApiUrl(postfix: string, language: Language): string {
@@ -112,34 +128,28 @@ export class TradeHttpService {
         return `${baseUrl}/trade/${postfix}`;
     }
 
-    private getAndTransform<TResponse>(url: string): Observable<TResponse> {
-        return this.http.get(url, {
-            responseType: 'text',
-            observe: 'response'
-        }).pipe(
-            retryWhen(errors => errors.pipe(
-                flatMap((error: HttpErrorResponse, count) => {
-                    if (count >= RETRY_COUNT) {
-                        return throwError(error);
-                    }
-                    if (error.status === 403) {
-                        return this.browser.retrieve(url).pipe(
-                            map(() => null)
-                        );
-                    }
-                    return of(null).pipe(delay(RETRY_DELAY));
-                })
-            )),
-            map(response => this.transformResponse(response))
-        );
+    private handleError(url: string, response: HttpErrorResponse, count: number): Observable<void> {
+        if (count >= RETRY_COUNT) {
+            return throwError(response);
+        }
+
+        switch (response.status) {
+            case 400:
+                try {
+                    const error = JSON.parse(response.error);
+                    const message = error?.error?.message || 'no message';
+                    const code = error?.error?.code || '-';
+                    return throwError(`${code}: ${message}`);
+                } catch{
+                    return throwError(response.error);
+                }
+            case 403:
+                return this.browser.retrieve(url);
+            case 429:
+                return of(null).pipe(delay(RETRY_LIMIT_DELAY));
+            default:
+                return of(null).pipe(delay(RETRY_DELAY));
+        }
     }
 
-    private transformResponse<TResponse>(response: HttpResponse<string>): TResponse {
-        const result = response.body.replace(
-            /\\u[\dA-Fa-f]{4}/g,
-            (match) => String.fromCharCode(
-                parseInt(match.replace(/\\u/g, ''), 16))
-        );
-        return JSON.parse(result) as TResponse;
-    }
 }