+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# Changelog
+
+## 0.6.14 (2020-03-26)
+
+- add nonunique rarity support (#477)
+- add changelog as tray entry and show after update (#471)
+- add display trade page error message (#468)
+- add renderer logging support log to file (#468)
+- update data to 3.10.1
+- fix auto update by calling quit and install even after normal quit (#474)
+- fix invisible values at map info (#472)
+
 ## 0.6.13 (2020-03-25)
 
 - add loading animations